+++ packages/toolbars-v3/web/src/ricos-toolbars.ts
@@ -1,6 +1,12 @@
 /* eslint-disable brace-style */
 import { TOOLBARS } from 'wix-rich-content-editor-common';
-import type { PublisherProvider, IRicosToolbar, IRicosToolbars, ToolbarType } from 'ricos-types';
+import type {
+  PublisherProvider,
+  IRicosToolbar,
+  IRicosToolbars,
+  ToolbarType,
+  KeyboardShortcut,
+} from 'ricos-types';
 
 type Topics = ['ricos.toolbars.functionality.buttonClick', 'ricos.toolbars.functionality.search'];
 
@@ -131,5 +137,33 @@ export class RicosToolbars implements IRicosToolbars {
       this.static,
     ];
     return toolbars.filter(toolbar => toolbar.type === type)[0];
+  }
+
+  getShortcuts(): KeyboardShortcut[] {
+    const focusShortcut: KeyboardShortcut = {
+      name: 'Focus Toolbar',
+      description: 'Focusing on open toolbar',
+      keys: { macOs: 'Ctrl+T', windows: 'Alt+T' },
+      command() {
+        const firstFloatingFormattingToolbarButton = document.querySelector(
+          `[data-hook='floating-formatting-toolbar'] [data-hook='toolbar-v3-visible'] Button`
+        ) as HTMLElement;
+        firstFloatingFormattingToolbarButton?.focus();
+
+        const firstStaticFormattingToolbarButton = document.querySelector(
+          `[data-hook='static-toolbar'] [data-hook='toolbar-v3-visible'] Button`
+        ) as HTMLElement;
+        firstStaticFormattingToolbarButton?.focus();
+
+        const firstLinkToolbarButton = document.querySelector(
+          `[data-hook='linkPluginToolbar'] [data-hook='toolbar-v3-visible'] a`
+        ) as HTMLElement;
+        firstLinkToolbarButton?.focus();
+      },
+      group: 'global',
+      enabled: true,
+    };
+
+    return [focusShortcut];
   }
 }