+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 
 - `toolbars-v3`
   - [#4328](https://github.com/wix/ricos/pull/4328) custom heading fonts fixes
+  - [#4287](https://github.com/wix/ricos/pull/4287) focus on/off formatting toolbar shortcut (tiptap)
 
 ## :house: Internal